app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile app/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
crispy-daisyui==0.7.0
    # via -r app/requirements.in
cryptography==45.0.4
    # via pyjwt
django==5.2.1
    # via
    #   -r app/requirements.in
    #   crispy-daisyui
    #   django-allauth
    #   django-cors-headers
    #   django-cotton
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-formtools
    #   django-reversion
    #   django-vite
    #   djangorestframework
django-allauth[socialaccount]==65.9.0
    # via -r app/requirements.in
django-cors-headers==4.7.0
    # via -r app/requirements.in
django-cotton==2.1.2
    # via -r app/requirements.in
django-crispy-forms==2.4
    # via
    #   -r app/requirements.in
    #   crispy-daisyui
django-debug-toolbar==5.2.0
    # via -r app/requirements.in
django-formtools==2.5.1
    # via -r app/requirements.in
django-ratelimit==4.1.0
    # via -r app/requirements.in
django-reversion==5.1.0
    # via -r app/requirements.in
django-vite==3.1.0
    # via -r app/requirements.in
djangorestframework==3.16.1
    # via -r app/requirements.in
exceptiongroup==1.3.0
    # via anyio
gunicorn==23.0.0
    # via -r app/requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via listmonk
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
listmonk==0.3.7
    # via -r app/requirements.in
markdown==3.8.2
    # via -r app/requirements.in
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via gunicorn
pillow==11.3.0
    # via -r app/requirements.in
psycopg2-binary==2.9.10
    # via -r app/requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via listmonk
pydantic-core==2.33.2
    # via pydantic
pyjwt[crypto]==2.10.1
    # via django-allauth
python-dotenv==1.1.1
    # via -r app/requirements.in
python-frontmatter==1.1.0
    # via -r app/requirements.in
pyyaml==6.0.2
    # via python-frontmatter
requests==2.32.4
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
strenum==0.4.15
    # via listmonk
typing-extensions==4.14.0
    # via
    #   anyio
    #   asgiref
    #   exceptiongroup
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
whitenoise==6.9.0
    # via -r app/requirements.in
